Add custom date formats
Adds custom date formats, created with a variable in config named 'datetimeformat', this variable is a string template, see https://strftime.org for more info. Quotes are now added around dates in the markdown metadata for consistent interpretation. New variable for rssitemtemplate and previewtemplate: fulldate. This gives the date in a format which can be used by <pubDate>